chore(deps): break award + campaign-context-panel cycles
Pattern 4: extract shared types to *.types.ts leaf files.

- src/components/award/awardIcons.tsx <-> awardIconComponents.tsx
  IconProps moved to awardIcons.types.ts; both halves import the leaf.
- src/components/campaign/ContextPanel.tsx <-> ContextPanelTabs.tsx
  ContextPanelMode + 5 data interfaces moved to ContextPanel.types.ts;
  re-exported from ContextPanel.tsx so existing public API is preserved
  (storybook + other consumers unaffected).

Verified: madge -> 0 circular deps in src/components/award and
src/components/campaign.
